+++ include/rmd/seed_matrix.cuh
@@ -47,7 +47,7 @@ public:
   void downloadB(float *host_align_row_maj) const;
   void downloadSumTempl(float *host_align_row_maj) const;
   void downloadConstTemplDenom(float *host_align_row_maj) const;
-  void downloadConvergence(int8_t *host_align_row_maj) const;
+  void downloadConvergence(int *host_align_row_maj) const;
   void downloadEpipolarMatches(float2 *host_align_row_maj) const;
   int getPatchSide() const { return dev_data_.patch.side; }
 #endif
@@ -61,7 +61,7 @@ private:
   // Measurement parameters
   DeviceImage<float> mu_, sigma_, a_, b_;
   // Convergence state
-  DeviceImage<int8_t> convergence_;
+  DeviceImage<int> convergence_;
   // Epipolar matches
   DeviceImage<float2> epipolar_matches_;
   DeviceData dev_data_;
