+++ include/rmd/texture_memory.cuh
@@ -15,7 +15,7 @@ texture<float, cudaTextureType2D, cudaReadModeElementType> sigma_tex;
 texture<float, cudaTextureType2D, cudaReadModeElementType> a_tex;
 texture<float, cudaTextureType2D, cudaReadModeElementType> b_tex;
 
-texture<int8_t, cudaTextureType2D, cudaReadModeElementType> convergence_tex;
+texture<int, cudaTextureType2D, cudaReadModeElementType> convergence_tex;
 texture<float2, cudaTextureType2D, cudaReadModeElementType> epipolar_matches_tex;
 
 template<typename ElementType>
